P2P2: Initialize peer PASN data and set P2P2 on P2P PASN auth frame RX
During PASN pairing verification, the PASN authentication frame is
received without the need for provision discovery. If the pairing
initiator's peer entry is removed and added again or the MAC address
is randomized, the Authentication frame is discarded because the PASN
data is not initialized. Therefore, ensure that the PASN data for the
peer is initialized if it is not already present. Also set the P2P2
bool variable to true.

Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>