mtk: hostapd: prevent NULL pointer access in hostapd_free_hapd_data()
In the hostapd_free_hapd_data(), other hapd on the same
hostapd_iface might already be freed so that iface->bss[j] is NULL.

For example, hostapd_setup_bss() fails on the first hapd and returns
the failure to hostapd_add_iface(). In the error handling of
hostapd_add_iface(), it goes through each hapd on the iface and does
the deinitialization, which including calling hostapd_free_hapd_data()
and setting iface->bss[i] to NULL. In other word, iface->bss[1] enters
it deinitialization with iface->bss[0] being NULL, and a NULL pointer
access happens in hostapd_free_hapd_data().

Signed-off-by: Michael-CY Lee <michael-cy.lee@mediatek.com>