PR: Initialize Proximity Ranging global context
Add changes to initialize and deinitialize the Proximity Ranging (PR)
global context and Makefile changes to enable the compilation of this
feature. The Proximity Ranging context will be global making it common
to all interfaces.

The compilation of changes related to Proximity Ranging can be
enabled using the CONFIG_PR flag.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>