PR: Add ranging capabilities in USD PR element
Create Proximity Ranging capability buffer that will be a part of a
larger Proximity Ranging element that is used to advertise the ranging
capabilities of the device. The capabilities are fetched from the global
proximity ranging context. The capabilities include name of the device,
PASN authentication support, EDCA ranging support, NTB ranging support,
and such.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>