PR: Add EDCA capabilities in USD PR element
Create a buffer that contains EDCA (802.11mc) based ranging capabilities
of a device which will be part of a Proximity Ranging element used to
advertise ranging capabilities of a device. These capabilities are added
to USD frames to enable their exchange with peer devices. The EDCA based
ranging capabilities have attributes such as the ability to act as an
EDCA based ranging initiator, the ability to act as a EDCA based ranging
responder, the channels on which the EDCA Ranging is supported by the
device and other device specific capabilities.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>