PR: Add NTB capabilities in USD PR element
Create a buffer that contains the NTB (802.11az) ranging capabilities of
a device. This buffer will be part of the Proximity Ranging Information
element used to advertise the device's ranging capabilities. These
capabilities are added to USD frames to enable their exchange with peer
devices.

The NTB-based ranging capabilities include attributes such as the
ability to act as an NTB ranging initiator, the ability to act as an NTB
ranging responder, the channels on which NTB ranging is supported by the
device, and other device-specific capabilities.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>