PR: Derive and add DIRA in USD PR element
Derive the Device Identity Resolution Attribute (DIRA) and add it to the
Proximity Ranging element. The DIRA attribute is used to resolve the
identity of a device even when MAC randomization causes the peer
device's MAC address to change. The DIRA attribute includes fields such
as the cipher version, a nonce, and a tag.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>