PR: Process EDCA Capability Attribute from USD PR element
Parse part of Proximity Ranging element corresponding to the Proximity
Ranging EDCA Capability Attribute and store the corresponding data in a
structure. This attribute contains capabilities that tell if the device
can act as an EDCA based ranging initiator, EDCA based ranging
responder, and a list of channels on which the device supports EDCA
based ranging and other device specific capabilities.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>