PR: Process 11az NTB Capability Attribute from USD PR element
Parse part of Proximity Ranging element corresponding to the Proximity
Ranging 11az NTB Capability Attribute and store the corresponding data
in a structure. This attribute contains capabilities that tell if the
device can act as an NTB ranging initiator or NTB ranging responder, a
list of channels on which the device supports NTB ranging, and other
device specific capabilities.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>