PR: Handle PR element in PASN Auth M2 and prepare M3
PASN Auth M2 frame is processed and then M3 is prepared based on the
operation mode received in M2 and capabilities received in M2 are also
validated. If an operation mode could not be decided or validation of
capabilities has failed, the status is set as failure and M3 is sent. If
PR capabilities are valid and operation mode received in M2 can be
honored, the status is set as success and operation mode is set in M3.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>