PR: Pass PR elements in USD frames for processing
Add changes needed to process Proximity Ranging attributes present in
USD frames. USD is performed with the PR attribute to exchange
capabilities specific to proximity ranging. The discovered device
capabilities along with the other details such as address and name are
stored in a list present within the proximity ranging global context.

This commit includes only the base framework for getting the information
into appropriate places. Actual processing of the elements will be added
in separate commits.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>