PR: Handle PR element in PASN Auth M1 and prepare M2
Validate the PR element received in PASN Auth M1. Attributes such as
capabilities and ranging capabilities are validated. In case where
validation of the PR element fails or processing of M1 frame fails or
operation mode could not be derived, status is set as failure and M2 is
sent. In case a PR elements could be validated and operation mode could
be derived, capabilities and preferred ranging type capabilities,
operation mode, and status as success are added to the PR element for M2
and sent.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>