P2P2: Enable P2P2 and PCC only for SME non-offload drivers by default
Enable P2P2 and PCC support by default for SME non-offload drivers.
Other drivers will enable support for P2P2 and PCC from the driver
capabilities.

This is needed to avoid enabling P2P2 functionality for SME-in-driver
cases where the driver is not actually capable of all the needed
functionality and there is no explicit driver capability indication to
determine this.

Signed-off-by: Vinay Gannevaram <quic_vganneva@quicinc.com>