mtk: hostapd: avoid clearing vht 160 support during downgrading bw
Avoid clearing BW160 support in VHT capabilities when downgrading bandwidth via channel switch.
Otherwise, the STA will store the incorrect VHT capabilities of its connected AP, which leads
to a VHT capabilities unsync issue.
In this case, the STA will not be able to follow the AP to switch from BW80 back to BW160,
since the VHT capabilities of the AP stored by the STA indicate that BW160 is not supported.

Signed-off-by: StanleyYP Wang <StanleyYP.Wang@mediatek.com>