MLD: Clear group keys for removed links
For AP link removal case, group keys (e.g., GTK, IGTK, BIGTK, etc.) were
not cleared for links that were removed in the NL80211_CMD_LINKS_REMOVED
event. Since the links are already removed, we should clear the keys
associated with these links for proper cleanup.

This functionality was missing in the current implementation, so address
this by clearing group keys for the removed links.

Signed-off-by: Kavita Kavita <quic_kkavita@quicinc.com>