supplicant: Do not attempt to scan frequencies on masked bands.
If the phy_bands option is selected, we should not be attempting to
find frequencies on masked bands. This fixes an issue where an interface
on a different sub-phy could cause this interface to not scan, as it
would be both restricted to the invalid band of the other interface.

Signed-off-by: Rory Little <rory@candelatech.com>