PASN: Reuse existing PTKSA cache keys instead of redoing PASN
When a successful PASN exchange has already occurred with a peer,
the derived keys are stored in the PTKSA cache. If another PASN
request is received for the same peer, these cached keys should
be reused. However, the cache was previously bypassed due to the
incorrect pairwise cipher argument.

Ensure the correct parameters are set in advance, from the RSNE data
allowing the cache to be used when appropriate.

Signed-off-by: Peddolla Harshavardhan Reddy <peddolla@qti.qualcomm.com>