supplicant: Improve scanning and sharing.
Share scan results with siblings if sibling uses subset
of frequencies of the requested scan, instead of 'equal'
comparison.

Use radio-wide min-scan-gap (per band) instead of per-sta
scan gap.  Stas share a single radio, so otherwise the min
scan gap doesn't really work.

Filter scan responses based on the phy-band for the
sta so that it cannot attempt to associate in the wrong
band.

Signed-off-by: Ben Greear <greearb@candelatech.com>