tests: Disable IPv6 in protected keep-alive test
When the new protected_keep_alive test was added, the @disable_ipv6
decorator was not moved. Disable IPv6 for both keep alive tests by
moving it to the run_ function.

Also remove the @remote_compatible decorator as it does not work with
IPv6 disablement. It was not removed in the patch to disable IPv6 in the
original test.

Fixes: fdf23679f823 ("tests: More coverage for WNM BSS max idle period management")
Fixes: 1bd5e35b8733 ("tests: Disable IPv6 in WNM keep-alive test")
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>