RRM: Don't request LCI and civic if not supported
Check that FTM responder capability bit in Extended Capabilities element
and LCI and Location Civic bits in RM Enabled capabilities element are
set before sending LCI and location civic request in neighbor report
request. If they are not set, don't include the corresponding
sub-elements in the neighbor report request. This change is according to
IEEE Std 802.11-2024, 11.10.10.2 (Requesting a neighbor report).

Signed-off-by: David Spinadel <david.spinadel@intel.com>
Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Peer, Ilan <ilan.peer@intel.com>