AP: Always verify and fix channel switch configuration
On the hostapd control interface, the CHAN_SWITCH command would do some
extra verification and also ensured that HT/VHT/HE/EHT is not being
disabled (as this is correctly supported currently). However, these
checks were missing on the wpa_supplicant side which would be used,
e.g., when a CHAN_SWITCH command is done on a GO interface.

Move these checks into the hostapd_parse_csa_settings() function to
ensure that they are correctly done in both cases. This fixes issues
where for example HT is getting accidentally disabled in a test, causing
a subsequent mac80211 disconnect. Some tests may occasionally trigger
this problem, but usually pass due to timing (this will be addressed by
a later commit).

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>