tests: Ensure chan_switch tests receive a beacon after switch
mac80211 might disconnect if the beacon is incorrect after the channel
switch due to a misconfiguration of the GO. This would happen in the
test as HT is not explicitly selected and was disabled after the switch.
The only reason the test passed was because the traffic check would
usually happen fast enough so that mac80211 had not yet processed the
first beacon.

Add a delay and a second check to ensure this error will be caught in
the future.

Signed-off-by: Benjamin Berg <benjamin.berg@intel.com>
Reviewed-by: Ilan Peer <ilan.peer@intel.com>
Reviewed-by: Andrei Otcheretianski <andrei.otcheretianski@intel.com>